backend/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy>=2.0.25,<3.0.0
alembic>=1.13.0,<2.0.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6
psutil>=5.9.0,<7.0.0
qrcode[pil]>=7.4,<9.0
aiogram>=3.4.0,<4.0.0
apscheduler>=3.10.0,<4.0.0
aiosmtplib>=3.0.0,<4.0.0
cryptography>=41.0.0,<45.0.0
aiofiles>=23.0.0,<25.0.0
httpx>=0.25.0,<1.0.0
